changed back to number

+++ Schemas/item_stock.sql
@@ -1,6 +1,6 @@
 # Stores item stock
 CREATE TABLE stock ( 
 itemid NUMBER(10),
-price DOUBLE(10,2),
+price NUMBER(10,2),
 PRIMARY KEY(itemid) 
 );
